Refactored using saga
