+++ package.json
@@ -15,6 +15,7 @@
     "react-scripts": "3.4.1",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
+    "redux-saga": "^1.1.3",
     "reselect": "^4.0.0"
   },
   "scripts": {
