+++ src/components/ApartList/ApartList.scss
@@ -1,55 +1,74 @@
 @import '../../styles/base/abstracts';
 .apart {
-	&__list {
-		display: grid;
-		grid-gap: rem(8);
-		//grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
-		grid-template-rows: rem(225) rem(225) rem(225);
-		grid-template-areas:
-			'a a a a a a b b b b b b'
-			'c c c d d d b b b b b b'
-			'e e e e e e f f f f g g';
-	}
-	&__item {
-		&:first-child {
-			grid-area: a;
-		}
-		&:nth-child(2) {
-			grid-area: b;
-		}
-		&:nth-child(3) {
-			grid-area: c;
-		}
-		&:nth-child(4) {
-			grid-area: d;
-		}
-		&:nth-child(5) {
-			grid-area: e;
-		}
-		&:nth-child(6) {
-			grid-area: f;
-		}
-		&:last-child {
-			grid-area: g;
-		}
-	}
-	&__view-all {
-		display: flex;
-		justify-content: center;
-		align-items: center;
+  &__list {
+    display: grid;
+    grid-gap: rem(8);
+    //grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
+    grid-template-rows: rem(225) rem(225) rem(225);
+    grid-template-areas:
+      'a a a a a a b b b b b b'
+      'c c c d d d b b b b b b'
+      'e e e e e e f f f f g g';
+  }
+  &__item {
+    animation: fade-in 0.5s ease-out;
+    animation-fill-mode: both;
+    &:first-child {
+      grid-area: a;
+    }
+    &:nth-child(2) {
+      grid-area: b;
+      animation-delay: 0.1s;
+    }
+    &:nth-child(3) {
+      grid-area: c;
+      animation-delay: 0.2s;
+    }
+    &:nth-child(4) {
+      grid-area: d;
+      animation-delay: 0.3s;
+    }
+    &:nth-child(5) {
+      grid-area: e;
+      animation-delay: 0.4s;
+    }
+    &:nth-child(6) {
+      grid-area: f;
+      animation-delay: 0.5s;
+    }
+    &:last-child {
+      grid-area: g;
+      animation-delay: 0.6s;
+    }
+  }
+  &__view-all {
+    display: flex;
+    justify-content: center;
+    align-items: center;
 
-		font-family: 'Source-Light', sans-serif;
+    font-family: 'Source-Light', sans-serif;
 
-		padding: rem(14);
+    padding: rem(14);
 
-		a {
-			color: $dark-blue;
-		}
-		svg {
-			margin-left: rem(4);
-		}
-		svg path {
-			fill: $dark-blue;
-		}
-	}
+    a {
+      color: $dark-blue;
+    }
+    svg {
+      margin-left: rem(4);
+    }
+    svg path {
+      fill: $dark-blue;
+    }
+  }
+}
+
+@keyframes fade-in {
+  0% {
+    opacity: 0;
+    transform: scale(0.5);
+  }
+  100% {
+    opacity: 1;
+    transform: scale(1);
+  }
 }
