+++ src/components/Button/Button.scss
@@ -14,7 +14,45 @@
 
 	cursor: pointer;
 
+	position: relative;
+
+	transform: translateY(0);
+
+	transition: all $anim-duratuion $anim-function;
+
 	&:focus {
 		outline: none;
+	}
+
+	&::after {
+		content: '';
+		display: block;
+		width: 100%;
+		height: 100%;
+
+		position: absolute;
+		top: 0;
+		left: 0;
+		z-index: -1;
+
+		background-color: rgba(184, 155, 114, 0.8);
+
+		transform: scaleX(0);
+		transform-origin: 0 50%;
+		transition: all $anim-duratuion $anim-function;
+	}
+
+	&:hover {
+		color: $dark-green;
+		box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.15);
+	}
+
+	&:active {
+		transform: translateY(2px);
+		box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.2);
+	}
+
+	&:hover::after {
+		transform: scaleX(1);
 	}
 }
