+++ src/store/aparts/index.js
@@ -2,3 +2,4 @@ export { default } from './reducer';
 export * from './selectors';
 export * from './actions';
 export * from './middlewares';
+export * from './saga';
